+++ src/Microsoft.Fast.Components.FluentUI/Components/List/FluentCombobox.razor.cs
@@ -60,4 +60,11 @@ public partial class FluentCombobox<TOption> : ListComponentBase<TOption>
         }
     }
 
+    protected override string? GetOptionValue(TOption? item)
+    {
+        if (item != null)
+            return OptionText.Invoke(item) ?? OptionValue.Invoke(item) ?? item.ToString();
+        else
+            return null;
+    }
 }